# docs/swagger.yaml
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.AgencyActiveDiscountItem:
    properties:
      company_name:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      discount_rate:
        type: number
      first_name:
        type: string
      last_name:
        type: string
    type: object
  dto.AgencyCustomerDiscountItem:
    properties:
      agency_share_with_tax:
        type: integer
      created_at:
        type: string
      discount_rate:
        type: number
      expires_at:
        type: string
      total_sent:
        type: integer
    type: object
  dto.AgencyCustomerItem:
    properties:
      company_name:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
    type: object
  dto.AgencyCustomerReportItem:
    properties:
      company_name:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      total_agency_share_with_tax:
        type: integer
      total_sent:
        type: integer
    type: object
  dto.AgencyCustomerReportResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AgencyCustomerReportItem'
        type: array
      message:
        type: string
      sum_total_agency_share_with_tax:
        type: integer
      sum_total_sent:
        type: integer
    type: object
...
  dto.CreateAgencyDiscountRequest:
    properties:
      customer_id:
        minimum: 1
        type: integer
      discount_rate:
        maximum: 0.5
        minimum: 0
        type: number
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - customer_id
    - discount_rate
    - name
    type: object
  dto.CreateAgencyDiscountResponse:
    properties:
      discount_rate:
        type: number
      message:
        type: string
    type: object
...
  dto.ListAgencyActiveDiscountsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AgencyActiveDiscountItem'
        type: array
      message:
        type: string
    type: object
  dto.ListAgencyCustomerDiscountsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AgencyCustomerDiscountItem'
        type: array
      message:
        type: string
    type: object
  dto.ListAgencyCustomersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AgencyCustomerItem'
        type: array
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/reports/agency/customers:
    get:
      consumes:
      - application/json
      description: Retrieve per-customer aggregated stats of campaigns for an agency
        with pagination and filters
      parameters:
      - description: Sorting (name_asc|name_desc|sent_desc|share_desc)
        in: query
        name: orderby
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Filter by customer name (first+last)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AgencyCustomerReportResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Get Agency Customer Report
      tags:
      - Reports
  /api/v1/reports/agency/customers/{customer_id}/discounts:
    get:
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListAgencyCustomerDiscountsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Customer Discounts Under Agency
      tags:
      - Reports
  /api/v1/reports/agency/customers/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListAgencyCustomersResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Agency Customers
      tags:
      - Reports
  /api/v1/reports/agency/discounts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create discount payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAgencyDiscountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateAgencyDiscountResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Create Agency Discount
      tags:
      - Reports
  /api/v1/reports/agency/discounts/active:
    get:
      parameters:
      - description: Filter by user name (first+last)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListAgencyActiveDiscountsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Active Discounts of Agency
      tags:
      - Reports
...
swagger: "2.0"